Add arsenal access limit
