+++ addons/retextures/CfgWFARArsenalAccess.hpp
@@ -1,0 +1,15 @@
+#define FINAL
+class cfgWFARArsenalAccess
+{
+	class IDX_VIRT_HEADGEAR
+	{
+         //Jews
+		76561198042318568[]={
+            #include "data\rhsusf\mich_covered\jews\weapons.hpp"
+        };
+        //Shark
+		76561198053950415[]={
+            #include "data\rhsusf\mich_covered\shark\weapons.hpp"
+        };
+	}
+};
