+++ addons/retextures/CfgWeapons.hpp
@@ -1,0 +1,17 @@
+class cfgWeapons {
+    class rhsusf_mich_helmet_marpatwd;
+    class rhsusf_mich_helmet_marpatwd_headset;
+
+    class rhsusf_mich_helmet_marpatwd_alt;
+    class rhsusf_mich_helmet_marpatwd_alt_headset;
+
+
+    class rhsusf_mich_helmet_marpatwd_norotos;
+    class rhsusf_mich_helmet_marpatwd_norotos_headset;
+
+    class rhsusf_mich_helmet_marpatwd_norotos_arc;
+    class rhsusf_mich_helmet_marpatwd_norotos_arc_headset;
+
+    #include "data\rhsusf\mich_covered\jews\CfgWeapons.hpp"
+    #include "data\rhsusf\mich_covered\shark\CfgWeapons.hpp"
+};
