+++ addons/retextures/functions/script_component.hpp
@@ -1,0 +1,1 @@
+#include "\z\wfar\addons\retextures\script_component.hpp"
